backend-python/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
httpx
yfinance
pandas
numpy
google-generativeai
openai
anthropic
